fixed up fs_setup

+++ installation/fs_setup.sh
@@ -1,31 +1,37 @@
 #!/bin/bash
 
 # check for a stable internet connection
-ping -c 1 -q google.com >&/dev/null; echo $?
+NET_CON=$(ping -c 1 -q google.com >&/dev/null; echo $?)
 
-# Set time
+if [$NET_CON != 0]; then
+    echo "No internet connection... Exiting.."
+    exit
+fi
+
+# set the time
 echo "Setting timezone"
 timedatectl set-ntp true
 
-# Read parition table
+# read the parition table
 
 sfdisk /dev/vda < ./vda.sfdisk
 
-# Format partitions
+# format partitions
 
 mkfs.ext4 /dev/vda2
 mkswap /dev/vda1
 swapon /dev/vda1
 
-# Mount filesystem
+# mount filesystems
 
 mount /dev/vda2 /mnt
 
-# Pacstrap
+# install necessary packages
+# TODO: Convert to csv file and read!
 
-yes ""  | pacstrap /mnt base base-devel linux linux-headers linux-firmware grub networkmanager network-manager-applet wpa_supplicant dialog os-prober mtools dosfstools intel-ucode xf86-video-intel mesa xorg gnome-shell gnome-terminal gdm apparmor vim
+yes ""  | pacstrap /mnt base base-devel linux linux-headers linux-firmware grub networkmanager network-manager-applet wpa_supplicant dialog os-prober mtools dosfstools intel-ucode xf86-video-intel mesa xorg gnome-shell gnome-terminal gdm apparmor vim firejail
 
-# fstab
+# generate the fstab file to boot
 
 genfstab -U /mnt >> /mnt/etc/fstab
